# cmd/docs/swagger.yaml
definitions:
  api.AccessTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  api.InternalServerErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
  api.InvalidCredentialsErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  api.InvalidRequestErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  api.LoginInUseErrorResponse:
    properties:
      error:
        example: login already in use
        type: string
    type: object
  api.RefreshTokenErrorResponse:
    properties:
      error:
        example: refresh token error
        type: string
    type: object
  api.ValidationErrorResponse:
    properties:
      error:
        example: validation error
        type: string
    type: object
  common.LoginRequest:
    properties:
      login:
        example: the_real_slim_shady
        maxLength: 50
        type: string
      password:
        example: password1234
        maxLength: 100
        type: string
    required:
    - login
    - password
    type: object
  common.RegisterRequest:
    properties:
      login:
        example: the_real_slim_shady
        maxLength: 50
        minLength: 2
        type: string
      name:
        example: Slim
        maxLength: 100
        type: string
      password:
        example: password1234
        maxLength: 100
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - employee
        - user
        example: admin
        type: string
      surname:
        example: Shady
        maxLength: 100
        type: string
    required:
    - login
    - name
    - password
    - role
    - surname
    type: object
host: localhost:8080
info:
  contact: {}
  description: Internal API for canteen web app
  license:
    name: Apache 2.0
  title: CanteenApp API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует существующего пользователя, устанавливает refresh
        токен в cookie и возвращает access токен в теле ответа.
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно аутентифицирован
          schema:
            $ref: '#/definitions/api.AccessTokenResponse'
        "400":
          description: Данные невалидны
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
        "401":
          description: Логин/пароль некорректен
          schema:
            $ref: '#/definitions/api.InvalidCredentialsErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.InternalServerErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/auth/logout:
    post:
      description: Инвалидирует refresh токен в cookie
      responses:
        "204":
          description: Успешный выход, тело ответа отсутствует
      summary: Выход из системы
      tags:
      - auth
  /api/auth/refresh:
    get:
      description: Проверяет refresh токен, установленный в cookie, и возврашает в
        теле ответа новый access токен
      produces:
      - application/json
      responses:
        "200":
          description: access токен успешно обновлен
          schema:
            $ref: '#/definitions/api.AccessTokenResponse'
        "401":
          description: Refresh токен не установлен или некорректен
          schema:
            $ref: '#/definitions/api.RefreshTokenErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.InternalServerErrorResponse'
      summary: Обновление access токена
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя, устанавливает refresh токен в cookie
        и возвращает access токен в теле ответа.
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/api.AccessTokenResponse'
        "400":
          description: Данные невалидны
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
        "409":
          description: Пользователь с таким логином уже существует
          schema:
            $ref: '#/definitions/api.LoginInUseErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.InternalServerErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
swagger: "2.0"